+++ tests/unit/Entities/UserTest.php
@@ -65,9 +65,6 @@ class UserTest extends TestCase
         $user = new User(['id' => 1, 'first_name' => 'John', 'last_name' => '`Taylor`']);
         self::assertEquals('John `Taylor`', $user->tryMention());
         self::assertEquals('John \`Taylor\`', $user->tryMention(true));
-
-        // Plain escapeMarkdown functionality.
-        self::assertEquals('a\`b\[c\*d\_e', $user->escapeMarkdown('a`b[c*d_e'));
     }
 
     public function testGetProperties()